Beginning of the original GUI project for Pixel Dust. Changing world slots from structures to class. Main project name reformulation, now being 'PixelDust.Game'. Initial construction of the world chunk system for better optimization and element processing. More consolidated division of files. Inclusion of files related to the chunk system throughout the project.

+++ src/PixelDust.Core/Elements/PElement.cs
@@ -7,10 +7,10 @@ namespace PixelDust.Core.Elements
         public string Name { get; protected set; }
         public string Description { get; protected set; }
         public string Category { get; protected set; }
-        public uint Id { get; internal set; }
+        public byte Id { get; internal set; }
         public Color Color { get; protected set; }
 
-        public double DefaultTemperature { get; protected set; } = 20f;
+        public float DefaultTemperature { get; protected set; } = 20f;
         public int DefaultDispersionRate { get; protected set; } = 1;
         public bool HasColorVariation { get; protected set; } = true;
